+++ projects/question/src/main/java/org/batfish/question/BgpAsnUseQuestionPlugin.java
@@ -25,14 +25,14 @@ public class BgpAsnUseQuestionPlugin extends QuestionPlugin {
 
     public static final String PROP_ASNS = "asns";
 
-    private SortedSetMultimap<Integer, String> _asns;
+    private SortedSetMultimap<Long, String> _asns;
 
-    public BgpAsnUseAnswerElement(@JsonProperty(PROP_ASNS) TreeMultimap<Integer, String> asns) {
+    public BgpAsnUseAnswerElement(@JsonProperty(PROP_ASNS) TreeMultimap<Long, String> asns) {
       _asns = (asns == null) ? TreeMultimap.create() : asns;
     }
 
     @JsonProperty(PROP_ASNS)
-    public SortedSetMultimap<Integer, String> getAsns() {
+    public SortedSetMultimap<Long, String> getAsns() {
       return _asns;
     }
 
@@ -40,7 +40,7 @@ public class BgpAsnUseQuestionPlugin extends QuestionPlugin {
     public String prettyPrint() {
       StringBuilder sb = new StringBuilder("Results for BGP ASN use\n");
       if (_asns != null) {
-        for (Integer asn : _asns.keySet()) {
+        for (Long asn : _asns.keySet()) {
           sb.append("  " + asn + "\n");
           for (String node : _asns.get(asn)) {
             sb.append("    " + node + "\n");
@@ -63,7 +63,7 @@ public class BgpAsnUseQuestionPlugin extends QuestionPlugin {
       BgpAsnUseQuestion question = (BgpAsnUseQuestion) _question;
 
       BgpAsnUseAnswerElement answerElement = new BgpAsnUseAnswerElement(null);
-      SortedSetMultimap<Integer, String> asns = TreeMultimap.create();
+      SortedSetMultimap<Long, String> asns = TreeMultimap.create();
       Map<String, Configuration> configurations = _batfish.loadConfigurations();
       Set<String> nodes = question.getNodeRegex().getMatchingNodes(_batfish);
       for (Entry<String, Configuration> e : configurations.entrySet()) {
@@ -82,7 +82,7 @@ public class BgpAsnUseQuestionPlugin extends QuestionPlugin {
         }
       }
       // is there streams way of multimap to multimap conversion with a filter?
-      for (Integer asn : asns.keySet()) {
+      for (Long asn : asns.keySet()) {
         if (asns.get(asn).size() >= question.getMinCount()) {
           answerElement.getAsns().putAll(asn, asns.get(asn));
         }
